yanwen: [train-preprocessor] sync with new extractor apis; fix bugs based on srl example;
